fix: put email helpers into own files to avoid nodemailer import
